YoloV3/PyTorch_YOLOv3-dena/requirements/requirements.txt
torch
cython
numpy
matplotlib
opencv_python
tensorboardX
PyYAML
pycocotools


